# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Финни — API справочников и прогресса",
    "description": "\nБэкенд игры «Питомец Финни».\n\n**Приложение работает без сервера.** Игра полностью офлайн: контент зашит в\nкод приложения, прогресс хранится на устройстве. Этот сервис необязательный —\nон нужен для двух вещей:\n\n1. **Справочники** (`/api/v1/content/*`) — тот же контент, что в приложении:\n   магазин, цели копилки, задания дорог «Математика» и «Финансы», питомцы и\n   правила экономики. JSON собирается из Dart-кода скриптом\n   `scripts/sync_from_mobile.py`. Разделы кэшируются по ETag\n   (`If-None-Match` → 304).\n2. **Прогресс** (`/api/v1/progress/*`) — резервная копия снимка локального\n   профиля (переустановка, смена устройства, проверка экспертом).\n\nПерсональные данные не собираются: профиль адресуется UUID, который\nгенерирует само приложение.\n",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/content/manifest": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Версии всех справочников",
        "description": "Лёгкий запрос «не пора ли обновить контент». Клиент хранит `content_version` и ETag каждого раздела и скачивает только изменившиеся.",
        "operationId": "get_manifest_api_v1_content_manifest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentManifest"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          }
        }
      }
    },
    "/api/v1/content/bundle": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Весь контент одним ответом",
        "description": "Для первого запуска и офлайн-кэша: каталог, цели, задания, термины, питомцы, экономика.",
        "operationId": "get_bundle_api_v1_content_bundle_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentBundle"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          }
        }
      }
    },
    "/api/v1/content/catalog": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Каталог покупок",
        "description": "Товары магазина приложения: цена, отдел, влияние на питомца (ТЗ 2.5.6) и правило «скидки дня».",
        "operationId": "get_catalog_api_v1_content_catalog_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "mandatory",
                    "optional",
                    "education"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "mandatory («надо») | optional («хочу») | education",
              "title": "Kind"
            },
            "description": "mandatory («надо») | optional («хочу») | education"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "food, hygiene, health, toys…",
              "title": "Category"
            },
            "description": "food, hygiene, health, toys…"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/catalog/{item_id}": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Одна позиция каталога",
        "operationId": "get_catalog_item_api_v1_content_catalog__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Идентификатор товара, например apple",
              "title": "Item Id"
            },
            "description": "Идентификатор товара, например apple"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/goals": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Цели копилки",
        "description": "Готовые цели и правила своей цели: границы суммы, картинки (ТЗ 2.5.7).",
        "operationId": "get_goals_api_v1_content_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalsFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          }
        }
      }
    },
    "/api/v1/content/goals/{goal_id}": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Одна цель",
        "operationId": "get_goal_api_v1_content_goals__goal_id__get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/quests": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Задания: дороги «Математика» и «Финансы»",
        "description": "Задания сгруппированы по дорогам и классам 1–4 (ТЗ 2.5.8). У каждого — ситуация, четыре варианта ответа кнопками, разбор и справка. Награда — монеты и опыт за первое верное решение; ошибка ничего не отнимает, а превращает задание в задачу «Повтори» (ТЗ 2.5.9).",
        "operationId": "get_quests_api_v1_content_quests_get",
        "parameters": [
          {
            "name": "track",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "math | finance",
              "title": "Track"
            },
            "description": "math | finance"
          },
          {
            "name": "grade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Класс 1–4",
              "title": "Grade"
            },
            "description": "Класс 1–4"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestsFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/quests/{quest_id}": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Одно задание",
        "operationId": "get_quest_api_v1_content_quests__quest_id__get",
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/glossary": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Справочник терминов",
        "description": "Короткие объяснения финансовых слов для справочного раздела (ТЗ 2.5.11).",
        "operationId": "get_glossary_api_v1_content_glossary_get",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "finance_1 | finance_2 | finance_3",
              "title": "Topic"
            },
            "description": "finance_1 | finance_2 | finance_3"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlossaryFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/pets": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Питомец: виды, окраски, этапы, эмоции, настроение",
        "description": "Три вида × три окраски = 9 комбинаций (ТЗ 2.6), перекраска, этапы взросления по уровню, правила настроения и готовые имена для кубика.",
        "operationId": "get_pets_api_v1_content_pets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetsFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          }
        }
      }
    },
    "/api/v1/content/economy": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Правила игровой экономики",
        "description": "Константы приложения: старт, доход от курсов, бонус за вход, опыт (с потолком за день) и уровни, расход показателей за день, вклад, огонёк с бонусами, план бюджета, обязательные правила. Блок `from_documents` — где приложение сознательно отличается от документов.",
        "operationId": "get_economy_api_v1_content_economy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomyFile"
                }
              }
            }
          },
          "304": {
            "description": "Контент не изменился — используйте локальную копию"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogFile": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date",
            "title": "Updated At"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Откуда в приложении взят контент"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CatalogCategory"
            },
            "type": "array",
            "minItems": 1,
            "title": "Categories"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            },
            "type": "array",
            "title": "Items"
          },
          "deal_of_day": {
            "$ref": "#/components/schemas/DealOfDay"
          },
          "purchase_rules": {
            "$ref": "#/components/schemas/PurchaseRules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "updated_at",
          "categories",
          "items",
          "deal_of_day",
          "purchase_rules"
        ],
        "title": "CatalogFile"
      },
      "CatalogItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 20,
            "title": "Title",
            "description": "До 20 символов — требование к карточке товара"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "kind": {
            "type": "string",
            "enum": [
              "mandatory",
              "optional",
              "education"
            ],
            "title": "Kind"
          },
          "price": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Price"
          },
          "emoji": {
            "type": "string",
            "title": "Emoji"
          },
          "effects": {
            "$ref": "#/components/schemas/ItemEffects"
          },
          "effect_text": {
            "type": "string",
            "title": "Effect Text",
            "description": "Реплика после использования предмета"
          },
          "badge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badge",
            "description": "Ярлык на карточке: «Хит», «Новинка»…"
          },
          "permanent": {
            "type": "boolean",
            "title": "Permanent",
            "default": false
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "common",
                  "rare",
                  "epic"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity",
            "description": "Редкость — только у украшений коллекции"
          },
          "income_after": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Income After",
            "description": "Доход за день после курса — только у обучения"
          },
          "resale_price": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Resale Price",
            "description": "За сколько магазин выкупит украшение (70 %)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "kind",
          "price",
          "emoji",
          "effects",
          "effect_text"
        ],
        "title": "CatalogItem"
      },
      "ContentBundle": {
        "properties": {
          "content_version": {
            "type": "string",
            "title": "Content Version"
          },
          "catalog": {
            "$ref": "#/components/schemas/CatalogFile"
          },
          "goals": {
            "$ref": "#/components/schemas/GoalsFile"
          },
          "quests": {
            "$ref": "#/components/schemas/QuestsFile"
          },
          "glossary": {
            "$ref": "#/components/schemas/GlossaryFile"
          },
          "pets": {
            "$ref": "#/components/schemas/PetsFile"
          },
          "economy": {
            "$ref": "#/components/schemas/EconomyFile"
          }
        },
        "type": "object",
        "required": [
          "content_version",
          "catalog",
          "goals",
          "quests",
          "glossary",
          "pets",
          "economy"
        ],
        "title": "ContentBundle",
        "description": "Весь контент одним ответом."
      },
      "ContentManifest": {
        "properties": {
          "content_version": {
            "type": "string",
            "title": "Content Version"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ContentSectionInfo"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "content_version",
          "sections"
        ],
        "title": "ContentManifest",
        "description": "Лёгкий ответ для проверки «не пора ли обновить локальный контент»."
      },
...
      "GlossaryFile": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date",
            "title": "Updated At"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Откуда в приложении взят контент"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GlossaryTerm"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "updated_at",
          "items"
        ],
        "title": "GlossaryFile"
      },
...
      "Goal": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "emoji": {
            "type": "string",
            "title": "Emoji"
          },
          "price": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Price"
          },
          "hint": {
            "type": "string",
            "title": "Hint",
            "description": "Совет Финни: сколько откладывать и как долго"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "emoji",
          "price",
          "hint"
        ],
        "title": "Goal"
      },
...
      "GoalsFile": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date",
            "title": "Updated At"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Откуда в приложении взят контент"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Goal"
            },
            "type": "array",
            "title": "Items"
          },
          "default_goal_id": {
            "type": "string",
            "title": "Default Goal Id"
          },
          "custom_goal": {
            "$ref": "#/components/schemas/CustomGoalRules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "updated_at",
          "items",
          "default_goal_id",
          "custom_goal"
        ],
        "title": "GoalsFile"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Lesson": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "track": {
            "type": "string",
            "title": "Track"
          },
          "grade": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Grade",
            "description": "Класс 1–4: сложность и награда"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "emoji": {
            "type": "string",
            "title": "Emoji"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 2,
            "title": "Options",
            "description": "Ответ выбирается кнопкой"
          },
          "correct_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Correct Index"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Разбор — показывается после ответа"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Справка: правило, которое стоит запомнить"
          },
          "reward": {
            "$ref": "#/components/schemas/LessonReward"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "track",
          "grade",
          "title",
          "emoji",
          "question",
          "options",
          "correct_index",
          "answer",
          "note",
          "reward"
        ],
        "title": "Lesson"
      },
...
      "QuestsFile": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date",
            "title": "Updated At"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Откуда в приложении взят контент"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/LessonTrack"
            },
            "type": "array",
            "title": "Tracks"
          },
          "rules": {
            "$ref": "#/components/schemas/LessonRules"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Lesson"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "updated_at",
          "tracks",
          "rules",
          "items"
        ],
        "title": "QuestsFile"
      },
...
    }
  },
...
}